requirements.txt
Flask
tensorflow
Pillow
matplotlib